Bytter til combobox for stønadstypevelger
