+++ frontend/src/routes/oppgave/$oppgaveid/-components/NySakAction.tsx
@@ -1,13 +1,13 @@
-import type {Journalpost, OppgaveMedSak, ProblemDetail} from "@generated";
-import {journalforNySakMutation} from "@generated/@tanstack/react-query.gen";
-import {Button, VStack} from "@navikt/ds-react";
-import {useMutation} from "@tanstack/react-query";
-import {useNavigate} from "@tanstack/react-router";
-import {useState} from "react";
-import {ErrorAlert} from "~/common/error/ErrorAlert";
-import type {StonadType} from "~/common/sak/sak.types";
-import type {FellesData} from "./JournalforForm";
-import {StonadsTypeVelger} from "./StonadsTypeVelger";
+import type { Journalpost, OppgaveMedSak, ProblemDetail } from "@generated";
+import { journalforNySakMutation } from "@generated/@tanstack/react-query.gen";
+import { Button, VStack } from "@navikt/ds-react";
+import { useMutation } from "@tanstack/react-query";
+import { useNavigate } from "@tanstack/react-router";
+import { useState } from "react";
+import { ErrorAlert } from "~/common/error/ErrorAlert";
+import type { StonadType } from "~/common/sak/sak.types";
+import { StonadsTypeVelger } from "../../../../common/sak/StonadsTypeVelger";
+import type { FellesData } from "./JournalforForm";
 
 interface Props {
     oppgaveMedSak: OppgaveMedSak;
@@ -48,7 +48,7 @@ export function NySakAction({ oppgaveMedSak, journalPost, defaultStonadstype, re
     return (
         <VStack gap="space-24">
             <StonadsTypeVelger
-                name="stonadstype"
+                label="Velg type stønad"
                 value={stonadstype}
                 error={error}
                 readOnly={readOnly}
