--- frontend/src/routes/oppgave/$oppgaveid/-components/StonadsTypeVelger.tsx
@@ -1,38 +1,0 @@
-import { Select } from "@navikt/ds-react";
-import { useSuspenseQuery } from "@tanstack/react-query";
-import { getKodeverkStonadsTypeOptions } from "~/common/sak/sak.query";
-import type { StonadType } from "~/common/sak/sak.types";
-
-interface Props {
-    value?: StonadType;
-    error?: string;
-    onChange: (type: StonadType | undefined) => void;
-    name: string;
-    readOnly?: boolean;
-}
-
-export function StonadsTypeVelger({ value, error, onChange, name, readOnly }: Props) {
-    const { data: stonadsTyper } = useSuspenseQuery(getKodeverkStonadsTypeOptions());
-    const selectValue: string = (value ?? "") as string;
-
-    return (
-        <Select
-            label="Velg type stønad"
-            name={name}
-            value={selectValue}
-            error={error}
-            readOnly={readOnly}
-            onChange={(e) => {
-                const v = e.target.value;
-                onChange(v === "" ? undefined : (v as StonadType));
-            }}
-        >
-            <option value={""}>-Velg type -</option>
-            {stonadsTyper.map((bt) => (
-                <option key={bt.type} value={bt.type}>
-                    {bt.navn}
-                </option>
-            ))}
-        </Select>
-    );
-}
